refactor(step2): 패키징
domain들 패키징
