+++ src/main/java/ladder/domain/Ladder.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 import ladder.drawable.RandomDraw;
