+++ src/main/java/ladder/domain/LadderGame.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 public class LadderGame {
     private Players players;
