+++ src/main/java/ladder/domain/Point.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 public class Point {
     private final int heightPosition;
