+++ src/test/java/ladder/VerticalLinesTest.java
@@ -1,5 +1,7 @@
 package ladder;
 
+import ladder.domain.VerticalLine;
+import ladder.domain.VerticalLines;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
